Move single-use fns to do_import and mark private
